demon/requirements.txt
torchvision
watermark
numpy
jupyter

